Reworked cropping.
It seems to have not worked as intended at all. Even the single test image gave me a wrong result.
I tried transcribing the ImageJ script as best as i could, but there are no comments and the functions are somewhat opaque are not documented clearly.
I added a blob filter that should make the result more robust, but it needs to be tested on more data.