Removed dependencies to os for path management and old glob2 library.
Using os for paths is deprecated and pathlib offers much cleaner syntax.
Cluster setup still needs testing, as this is a major change that might break something or an error with path and string usage might have slipped through.

Glob2 can be dropped, as all functionality is contained much nicer in pathlib's glob.